+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/cheatcodes/precalculate_address.rs
@@ -1,13 +1,11 @@
 use crate::CheatnetState;
-use blockifier::execution::execution_utils::felt_to_stark_felt;
 use cairo_felt::Felt252;
 use conversions::IntoConv;
 use starknet::core::types::FieldElement;
 use starknet_api::core::{calculate_contract_address, ClassHash, ContractAddress};
-use starknet_api::hash::StarkFelt;
-use starknet_api::transaction::Calldata;
 
 use crate::constants as crate_constants;
+use crate::runtime_extensions::common::create_execute_calldata;
 
 impl CheatnetState {
     #[must_use]
@@ -28,9 +26,4 @@ impl CheatnetState {
         )
         .unwrap()
     }
-}
-
-pub fn create_execute_calldata(calldata: &[Felt252]) -> Calldata {
-    let calldata: Vec<StarkFelt> = calldata.iter().map(felt_to_stark_felt).collect();
-    Calldata(calldata.into())
 }
