+++ crates/cheatnet/src/runtime_extensions/mod.rs
@@ -1,4 +1,5 @@
 pub mod call_to_blockifier_runtime_extension;
 pub mod cheatable_starknet_runtime_extension;
+pub mod common;
 pub mod forge_runtime_extension;
 pub mod io_runtime_extension;
